+++ exo-go/src/aws/setup_for_terraform.go
