+++ exo-go/src/deploy/deployer.go
@@ -1,4 +1,4 @@
-package application
+package deploy
 
 import (
 	"fmt"
